moved the data fetching into the ProfileDaos component so that it doesnt block rendering on the profile
